# projects/shop-app/docs/apidocs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "基于 OneX 技术栈的电商后端服务，提供用户登录注册等功能。",
        "title": "Shop App API",
        "contact": {
            "name": "mungdong"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:5555",
    "basePath": "/api",
    "paths": {
        "/api/config": {
            "get": {
                "description": "返回前端可用的全局配置（如默认语言）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统"
                ],
                "summary": "获取全局配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ConfigResponse"
                        }
                    }
                }
            }
        },
        "/api/healthz": {
            "get": {
                "description": "探测服务是否存活",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统"
                ],
                "summary": "健康检查",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.HealthzResponse"
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "description": "通过用户名密码登录，返回 JWT Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.LoginResponse"
                        }
                    }
                }
            }
        },
        "/api/refresh-token": {
            "put": {
                "description": "使用当前有效的 Token 刷新获取新 Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "刷新令牌",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.RefreshTokenResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users": {
            "get": {
                "description": "分页查询用户列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "查询用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "default": 0,
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "default": 10,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ListUserResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新用户（注册），无需认证",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "注册用户",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.CreateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.CreateUserResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{userID}": {
            "get": {
                "description": "获取指定用户的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "查询用户详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户 ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.GetUserResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "更新指定用户的信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "更新用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户 ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新用户请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.UpdateUserResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户 ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.DeleteUserResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{userID}/change-password": {
            "put": {
                "description": "修改指定用户的密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户 ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改密码请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ChangePasswordResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_onexstack_shop-app_pkg_api_apiserver_v1.ServiceStatus": {
            "type": "integer",
            "format": "int32",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "ServiceStatus_Healthy",
                "ServiceStatus_Unhealthy"
            ]
        },
        "handler.ConfigResponse": {
            "type": "object",
            "properties": {
                "defaultLanguage": {
                    "description": "DefaultLanguage 默认语言.",
                    "type": "string"
                }
            }
        },
        "timestamppb.Timestamp": {
            "type": "object",
            "properties": {
                "nanos": {
                    "description": "Non-negative fractions of a second at nanosecond resolution. This field is\nthe nanosecond portion of the duration, not an alternative to seconds.\nNegative second values with fractions must still have non-negative nanos\nvalues that count forward in time. Must be between 0 and 999,999,999\ninclusive.",
                    "type": "integer"
                },
                "seconds": {
                    "description": "Represents seconds of UTC time since Unix epoch 1970-01-01T00:00:00Z. Must\nbe between -315576000000 and 315576000000 inclusive (which corresponds to\n0001-01-01T00:00:00Z to 9999-12-31T23:59:59Z).",
                    "type": "integer"
                }
            }
        },
        "v1.ChangePasswordRequest": {
            "type": "object",
            "properties": {
                "newPassword": {
                    "description": "newPassword 表示准备修改的新密码",
                    "type": "string"
                },
                "oldPassword": {
                    "description": "oldPassword 表示当前密码",
                    "type": "string"
                },
                "userID": {
                    "description": "userID 表示用户 ID",
                    "type": "string"
                }
            }
        },
        "v1.ChangePasswordResponse": {
            "type": "object"
        },
        "v1.CreateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "email 表示用户电子邮箱",
                    "type": "string"
                },
                "nickname": {
                    "description": "nickname 表示用户昵称",
                    "type": "string"
                },
                "password": {
                    "description": "password 表示用户密码",
                    "type": "string"
                },
                "phone": {
                    "description": "phone 表示用户手机号",
                    "type": "string"
                },
                "username": {
                    "description": "username 表示用户名称",
                    "type": "string"
                }
            }
        },
        "v1.CreateUserResponse": {
            "type": "object",
            "properties": {
                "userID": {
                    "description": "userID 表示新创建的用户 ID",
                    "type": "string"
                }
            }
        },
        "v1.DeleteUserResponse": {
            "type": "object"
        },
        "v1.GetUserResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "description": "user 表示返回的用户信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/v1.User"
                        }
                    ]
                }
            }
        },
        "v1.HealthzResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "message represents an optional status message, providing additional\ninformation about the service's health.",
                    "type": "string"
                },
                "status": {
                    "description": "status represents the health status of the service.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_onexstack_shop-app_pkg_api_apiserver_v1.ServiceStatus"
                        }
                    ]
                },
                "timestamp": {
                    "description": "timestamp represents the timestamp of the request.",
                    "type": "string"
                }
            }
        },
        "v1.ListUserResponse": {
            "type": "object",
            "properties": {
                "totalCount": {
                    "description": "totalCount 表示总用户数",
                    "type": "integer"
                },
                "users": {
                    "description": "users 表示用户列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.User"
                    }
                }
            }
        },
        "v1.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "password 表示用户密码",
                    "type": "string"
                },
                "username": {
                    "description": "username 表示用户名称",
                    "type": "string"
                }
            }
        },
        "v1.LoginResponse": {
            "type": "object",
            "properties": {
                "expireAt": {
                    "description": "expireAt 表示该 token 的过期时间",
                    "allOf": [
                        {
                            "$ref": "#/definitions/timestamppb.Timestamp"
                        }
                    ]
                },
                "token": {
                    "description": "token 表示返回的身份验证令牌",
                    "type": "string"
                }
            }
        },
        "v1.RefreshTokenResponse": {
            "type": "object",
            "properties": {
                "expireAt": {
                    "description": "expireAt 表示该 token 的过期时间",
                    "allOf": [
                        {
                            "$ref": "#/definitions/timestamppb.Timestamp"
                        }
                    ]
                },
                "token": {
                    "description": "token 表示返回的身份验证令牌",
                    "type": "string"
                }
            }
        },
        "v1.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "email 表示可选的用户电子邮箱",
                    "type": "string"
                },
                "nickname": {
                    "description": "nickname 表示可选的用户昵称",
                    "type": "string"
                },
                "phone": {
                    "description": "phone 表示可选的用户手机号",
                    "type": "string"
                },
                "userID": {
                    "description": "userID 表示用户 ID",
                    "type": "string"
                },
                "username": {
                    "description": "username 表示可选的用户名称",
                    "type": "string"
                }
            }
        },
        "v1.UpdateUserResponse": {
            "type": "object"
        },
        "v1.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "createdAt 表示用户注册时间",
                    "allOf": [
                        {
                            "$ref": "#/definitions/timestamppb.Timestamp"
                        }
                    ]
                },
                "email": {
                    "description": "email 表示用户电子邮箱",
                    "type": "string"
                },
                "nickname": {
                    "description": "nickname 表示用户昵称",
                    "type": "string"
                },
                "phone": {
                    "description": "phone 表示用户手机号",
                    "type": "string"
                },
                "postCount": {
                    "description": "postCount 表示用户拥有的博客数量",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "updatedAt 表示用户最后更新时间",
                    "allOf": [
                        {
                            "$ref": "#/definitions/timestamppb.Timestamp"
                        }
                    ]
                },
                "userID": {
                    "description": "userID 表示用户 ID",
                    "type": "string"
                },
                "username": {
                    "description": "username 表示用户名称",
                    "type": "string"
                }
            }
        }
    }
}