# projects/shop-app/pkg/api/apiserver/v1/healthz.pb.go
// Healthz API definition, which includes health check response messages and status.

// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.36.11
// 	protoc        v4.23.4
// source: apiserver/v1/healthz.proto

package v1

import (
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	reflect "reflect"
	sync "sync"
	unsafe "unsafe"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

// ServiceStatus represents the health status of the service.
type ServiceStatus int32

const (
	// Healthy indicates that the service is healthy.
	ServiceStatus_Healthy ServiceStatus = 0
	// Unhealthy indicates that the service is unhealthy.
	ServiceStatus_Unhealthy ServiceStatus = 1
)

// Enum value maps for ServiceStatus.
var (
	ServiceStatus_name = map[int32]string{
		0: "Healthy",
		1: "Unhealthy",
	}
	ServiceStatus_value = map[string]int32{
		"Healthy":   0,
		"Unhealthy": 1,
	}
)

func (x ServiceStatus) Enum() *ServiceStatus {
	p := new(ServiceStatus)
	*p = x
	return p
}

func (x ServiceStatus) String() string {
	return protoimpl.X.EnumStringOf(x.Descriptor(), protoreflect.EnumNumber(x))
}

func (ServiceStatus) Descriptor() protoreflect.EnumDescriptor {
	return file_apiserver_v1_healthz_proto_enumTypes[0].Descriptor()
}

func (ServiceStatus) Type() protoreflect.EnumType {
	return &file_apiserver_v1_healthz_proto_enumTypes[0]
}

func (x ServiceStatus) Number() protoreflect.EnumNumber {
	return protoreflect.EnumNumber(x)
}

// Deprecated: Use ServiceStatus.Descriptor instead.
func (ServiceStatus) EnumDescriptor() ([]byte, []int) {
	return file_apiserver_v1_healthz_proto_rawDescGZIP(), []int{0}
}

// HealthzResponse represents the response structure for a health check.
type HealthzResponse struct {
	state protoimpl.MessageState `protogen:"open.v1"`
	// status represents the health status of the service.
	Status ServiceStatus `protobuf:"varint,1,opt,name=status,proto3,enum=apiserver.v1.ServiceStatus" json:"status,omitempty"`
	// timestamp represents the timestamp of the request.
	Timestamp string `protobuf:"bytes,2,opt,name=timestamp,proto3" json:"timestamp,omitempty"`
	// message represents an optional status message, providing additional
	// information about the service's health.
	Message       string `protobuf:"bytes,3,opt,name=message,proto3" json:"message,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *HealthzResponse) Reset() {
	*x = HealthzResponse{}
	mi := &file_apiserver_v1_healthz_proto_msgTypes[0]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *HealthzResponse) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*HealthzResponse) ProtoMessage() {}

func (x *HealthzResponse) ProtoReflect() protoreflect.Message {
	mi := &file_apiserver_v1_healthz_proto_msgTypes[0]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use HealthzResponse.ProtoReflect.Descriptor instead.
func (*HealthzResponse) Descriptor() ([]byte, []int) {
	return file_apiserver_v1_healthz_proto_rawDescGZIP(), []int{0}
}

func (x *HealthzResponse) GetStatus() ServiceStatus {
	if x != nil {
		return x.Status
	}
	return ServiceStatus_Healthy
}

func (x *HealthzResponse) GetTimestamp() string {
	if x != nil {
		return x.Timestamp
	}
	return ""
}

func (x *HealthzResponse) GetMessage() string {
	if x != nil {
		return x.Message
	}
	return ""
}

var File_apiserver_v1_healthz_proto protoreflect.FileDescriptor

const file_apiserver_v1_healthz_proto_rawDesc = "" +
	"\n" +
	"\x1aapiserver/v1/healthz.proto\x12\fapiserver.v1\"~\n" +
	"\x0fHealthzResponse\x123\n" +
	"\x06status\x18\x01 \x01(\x0e2\x1b.apiserver.v1.ServiceStatusR\x06status\x12\x1c\n" +
	"\ttimestamp\x18\x02 \x01(\tR\ttimestamp\x12\x18\n" +
	"\amessage\x18\x03 \x01(\tR\amessage*+\n" +
	"\rServiceStatus\x12\v\n" +
	"\aHealthy\x10\x00\x12\r\n" +
	"\tUnhealthy\x10\x01B7Z5github.com/onexstack/shop-app/pkg/api/apiserver/v1;v1b\x06proto3"

var (
	file_apiserver_v1_healthz_proto_rawDescOnce sync.Once
	file_apiserver_v1_healthz_proto_rawDescData []byte
)

func file_apiserver_v1_healthz_proto_rawDescGZIP() []byte {
	file_apiserver_v1_healthz_proto_rawDescOnce.Do(func() {
		file_apiserver_v1_healthz_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_apiserver_v1_healthz_proto_rawDesc), len(file_apiserver_v1_healthz_proto_rawDesc)))
	})
	return file_apiserver_v1_healthz_proto_rawDescData
}

var file_apiserver_v1_healthz_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
var file_apiserver_v1_healthz_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
var file_apiserver_v1_healthz_proto_goTypes = []any{
	(ServiceStatus)(0),      // 0: apiserver.v1.ServiceStatus
	(*HealthzResponse)(nil), // 1: apiserver.v1.HealthzResponse
}
var file_apiserver_v1_healthz_proto_depIdxs = []int32{
	0, // 0: apiserver.v1.HealthzResponse.status:type_name -> apiserver.v1.ServiceStatus
	1, // [1:1] is the sub-list for method output_type
	1, // [1:1] is the sub-list for method input_type
	1, // [1:1] is the sub-list for extension type_name
	1, // [1:1] is the sub-list for extension extendee
	0, // [0:1] is the sub-list for field type_name
}

func init() { file_apiserver_v1_healthz_proto_init() }
func file_apiserver_v1_healthz_proto_init() {
	if File_apiserver_v1_healthz_proto != nil {
		return
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: unsafe.Slice(unsafe.StringData(file_apiserver_v1_healthz_proto_rawDesc), len(file_apiserver_v1_healthz_proto_rawDesc)),
			NumEnums:      1,
			NumMessages:   1,
			NumExtensions: 0,
			NumServices:   0,
		},
		GoTypes:           file_apiserver_v1_healthz_proto_goTypes,
		DependencyIndexes: file_apiserver_v1_healthz_proto_depIdxs,
		EnumInfos:         file_apiserver_v1_healthz_proto_enumTypes,
		MessageInfos:      file_apiserver_v1_healthz_proto_msgTypes,
	}.Build()
	File_apiserver_v1_healthz_proto = out.File
	file_apiserver_v1_healthz_proto_goTypes = nil
	file_apiserver_v1_healthz_proto_depIdxs = nil
}